feat: project detail

+++ server/routes/api.php
@@ -29,13 +29,15 @@ Route::post('/auth/verify', [AuthController::class, 'verify']);
 Route::post('/auth/reset', [AuthController::class, 'reset']);
 Route::middleware('jwt.auth')->get('/auth/token', [AuthController::class, 'show']);
 
-Route::middleware('jwt.auth', 'role:admin')->get('/role', [RoleController::class, 'index']);
+Route::middleware('jwt.auth', 'role:superadmin,admin')->get('/role', [RoleController::class, 'index']);
 
-Route::middleware('jwt.auth', 'role:admin')->get('/user', [UserController::class, 'index']);
+Route::middleware('jwt.auth', 'role:superadmin,admin')->get('/user', [UserController::class, 'index']);
+Route::middleware('jwt.auth', 'role:superadmin,admin')->get('/user/export', [UserController::class, 'export']);
 Route::middleware('jwt.auth')->get('/user/{userId}', [UserController::class, 'show']);
-Route::middleware('jwt.auth')->put('/user/{userId}', [UserController::class, 'update']);
-Route::middleware('jwt.auth', 'role:admin')->put('/user/{userId}/role', [UserController::class, 'update_role']);
-Route::middleware('jwt.auth', 'role:admin')->delete('/user/{userId}', [UserController::class, 'destroy']);
+Route::middleware('jwt.auth')->post('/user/{userId}', [UserController::class, 'update']);
+Route::middleware('jwt.auth', 'role:superadmin,admin')->put('/user/{userId}/role', [UserController::class, 'update_role']);
+Route::middleware('jwt.auth')->put('/user/{userId}/password', [UserController::class, 'update_password']);
+Route::middleware('jwt.auth', 'role:superadmin,admin')->delete('/user/{userId}', [UserController::class, 'destroy']);
 
 Route::middleware('jwt.auth')->get('/status', [StatusController::class, 'index']);
 
@@ -54,7 +56,9 @@ Route::middleware('jwt.auth', 'role:admin')->post('/project-user', [ProjectUserC
 
 Route::middleware('jwt.auth')->get('/issue', [IssueController::class, 'index']);
 Route::middleware('jwt.auth')->get('/issue/{issueId}', [IssueController::class, 'show']);
+Route::middleware('jwt.auth')->get('/issue/status/{statusId}', [IssueController::class, 'getIssueByStatus']);
 Route::middleware('jwt.auth')->post('/issue', [IssueController::class, 'store']);
+Route::middleware('jwt.auth')->post('/issue/{issueId}/status', [IssueController::class, 'changeIssueStatus']);
 
 Route::middleware('jwt.auth')->get('/comment', [CommentController::class, 'index']);
 Route::middleware('jwt.auth')->post('/comment', [CommentController::class, 'store']);
